Refactor OAuth flow to use proper API routes
- Move Google OAuth redirect to /api/auth/google route
- Create separate callback handler at /api/auth/google/callback
- Remove server action approach for cleaner architecture
- Update login button to redirect to fixed /api/auth/google route
- Update README with correct callback URI configuration
- Simplify client-side logic by removing environment variable dependencies
